bpf_test: Add crash verify and prog load mode
If we trigger a crash, we need to quickly identify which prog leads to
this. Add a crash verify mode, it scans each proc's directory and loads
the last prog. If the kernel crashes again, we found the cause.

If we found a false positive, we need to load the target prog again but
with detailed verifier log information and potentially execution info.
Add a pure load mode for this.

Signed-off-by: Hao Sun <hao.sun@inf.ethz.ch>